files: upload frontend file on server
