--- server/public/assets/app-auth-DYLPzh_K.js
@@ -1,5 +1,0 @@
-var re=Object.defineProperty,ae=Object.defineProperties;var oe=Object.getOwnPropertyDescriptors;var U=Object.getOwnPropertySymbols;var ne=Object.prototype.hasOwnProperty,ie=Object.prototype.propertyIsEnumerable;var G=(s,t,r)=>t in s?re(s,t,{enumerable:!0,configurable:!0,writable:!0,value:r}):s[t]=r,x=(s,t)=>{for(var r in t||(t={}))ne.call(t,r)&&G(s,r,t[r]);if(U)for(var r of U(t))ie.call(t,r)&&G(s,r,t[r]);return s},z=(s,t)=>ae(s,oe(t));var j=(s,t,r)=>new Promise((d,o)=>{var u=l=>{try{m(r.next(l))}catch(c){o(c)}},n=l=>{try{m(r.throw(l))}catch(c){o(c)}},m=l=>l.done?d(l.value):Promise.resolve(l.value).then(u,n);m((r=r.apply(s,t)).next())});import{r as g,u as v,j as e,F as le,a as ce,b as de,L as E,H as me,c as J,k as O,d as D,e as L,f as ue,G as ge,g as W,h as we,i as he,l as pe,m as xe}from"./vendor-react-RXQ05gJ2.js";import{P as fe,C as je,A as ye,S as q,F as T,E as K,a as I,B as X,b as R,c as N,d as Y}from"./app-components-DHLBCYYN.js";import{P as C,T as be,M as ke,C as V,a as Se,S as ve,b as ze,W as Ne,G as Ee,A as Pe,z as f,t as F}from"./vendor-other-C1_ffh8T.js";import{m as k,A as Me}from"./vendor-framer-motion-SZOJYst1.js";import{h as P}from"./app-services-tJo8l8LQ.js";const Ae="modulepreload",Ie=function(s){return"/"+s},Z={},B=function(t,r,d){let o=Promise.resolve();if(r&&r.length>0){document.getElementsByTagName("link");const n=document.querySelector("meta[property=csp-nonce]"),m=(n==null?void 0:n.nonce)||(n==null?void 0:n.getAttribute("nonce"));o=Promise.allSettled(r.map(l=>{if(l=Ie(l),l in Z)return;Z[l]=!0;const c=l.endsWith(".css"),i=c?'[rel="stylesheet"]':"";if(document.querySelector(`link[href="${l}"]${i}`))return;const a=document.createElement("link");if(a.rel=c?"stylesheet":Ae,c||(a.as="script"),a.crossOrigin="",a.href=l,m&&a.setAttribute("nonce",m),document.head.appendChild(a),c)return new Promise((w,h)=>{a.addEventListener("load",w),a.addEventListener("error",()=>h(new Error(`Unable to preload CSS for ${l}`)))})}))}function u(n){const m=new Event("vite:preloadError",{cancelable:!0});if(m.payload=n,window.dispatchEvent(m),!m.defaultPrevented)throw n}return o.then(n=>{for(const m of n||[])m.status==="rejected"&&u(m.reason);return t().catch(u)})},Le="https://codelinesjs.pl/api/",Fe=()=>{const[s,t]=g.useState(!1),[r,d]=g.useState(null),[o,u]=g.useState(!1),[n,m]=g.useState(!0),[l,c]=g.useState(null);return{loading:s,setLoading:t,error:r,setError:d,isAuthenticated:o,setIsAuthenticated:u,isLoading:n,setIsLoading:m,user:l,setUser:c}},Ce=s=>{const t=v(),{setLoading:r,setError:d,setIsAuthenticated:o,setUser:u}=s;return(m,l,c=!1)=>j(void 0,null,function*(){try{r(!0),d(null);const i=yield P.post("auth/login",{email:m,password:l,rememberMe:c});if(i.error)throw new Error(i.error);if(!i.data)throw new Error("Nieznany błąd logowania. Spróbuj ponownie później.");const{token:a,user:w}=i.data;localStorage.removeItem("token"),sessionStorage.removeItem("token"),c?localStorage.setItem("token",a):sessionStorage.setItem("token",a),u(w),o(!0),t("/dashboard")}catch(i){const a=i instanceof Error?i.message:"Wystąpił błąd podczas logowania";d(a),o(!1)}finally{r(!1)}})},_e=s=>{const t=v(),{setLoading:r,setError:d,setIsAuthenticated:o,setUser:u}=s;return(m,l,c)=>j(void 0,null,function*(){try{r(!0),d(null);const i=yield P.post("auth/register",{email:m,password:l,username:c});if(i.error)throw i.error.includes("już istnieje")?(d("Użytkownik o podanym adresie email lub nazwie użytkownika już istnieje w systemie"),new Error("Użytkownik o podanym adresie email lub nazwie użytkownika już istnieje w systemie")):(d(i.error),new Error(i.error));if(!i.data){const h="Nieznany błąd rejestracji. Spróbuj ponownie później.";throw d(h),new Error(h)}const{token:a,user:w}=i.data;localStorage.removeItem("token"),sessionStorage.removeItem("token"),sessionStorage.setItem("token",a),u&&w&&u(w),o(!0),t("/dashboard")}catch(i){const a=i instanceof Error?i.message:"Wystąpił błąd podczas rejestracji";d(a),o(!1)}finally{r(!1)}})},Re=s=>{const{setLoading:t,setError:r}=s;return o=>j(void 0,null,function*(){var u;try{t(!0),r(null);const n=yield P.post("auth/forgot-password",{email:o});if(n.error)throw new Error(n.error);return((u=n.data)==null?void 0:u.message)||"Link do resetowania hasła został wysłany na Twój adres email."}catch(n){if(n instanceof Error)throw r(n.message),n;{const m="Wystąpił błąd podczas wysyłania linku resetującego hasło";throw r(m),new Error(m)}}finally{t(!1)}})},We=s=>{const t=v(),{setLoading:r,setIsAuthenticated:d,setUser:o}=s;return()=>{try{r(!0),localStorage.removeItem("token"),sessionStorage.removeItem("token"),o&&o(null),d(!1),t("/")}finally{r(!1)}}},Te=s=>{const t=v(),{setLoading:r,setError:d,setIsAuthenticated:o,setUser:u}=s;return(m,l=!1)=>j(void 0,null,function*(){try{if(r(!0),d(null),!m.credential){d("Nie udało się uzyskać tokenu uwierzytelniającego z Google. Spróbuj ponownie."),o(!1);return}const c=yield P.post("auth/google-login",{credential:m.credential,rememberMe:l});if(c.error)throw new Error(c.error);if(!c.data)throw new Error("Nieznany błąd logowania przez Google. Spróbuj ponownie później.");const i=c.data,a=i.token,w=i.user||i.user&&i.user.user;l?localStorage.setItem("token",a):sessionStorage.setItem("token",a),u&&u(w),o(!0),t("/dashboard")}catch(c){const i=c instanceof Error?c.message:"Wystąpił błąd podczas logowania przez Google";d(i),o(!1)}finally{r(!1)}})},Be=s=>{const t=v(),{setLoading:r,setError:d}=s;return(u,n,m)=>j(void 0,null,function*(){var l;try{if(!u||u.length<10)throw new Error("Nieprawidłowy token resetowania hasła. Sprawdź, czy link jest poprawny.");if(n!==m)throw new Error("Hasła nie są identyczne. Upewnij się, że oba pola zawierają to samo hasło.");r(!0),d(null);const c=yield P.post("auth/reset-password",{token:u,password:n,confirmPassword:m});if(c.error)throw new Error(c.error);return setTimeout(()=>{t("/logowanie")},3e3),((l=c.data)==null?void 0:l.message)||"Hasło zostało pomyślnie zmienione. Za chwilę zostaniesz przekierowany do strony logowania."}catch(c){if(c instanceof Error)throw d(c.message),c;{const i="Wystąpił nieznany błąd podczas resetowania hasła";throw d(i),new Error(i)}}finally{r(!1)}})},He=s=>{const t=Ce(s),r=_e(s),d=Re(s),o=We(s),u=Te(s),n=Be(s);return{login:t,register:r,forgotPassword:d,logout:o,loginWithGoogle:u,resetPassword:n}},M=()=>{const s=Fe(),t=He(s),r=v(),[d,o]=g.useState(!0),u=()=>j(void 0,null,function*(){try{o(!0);const l=localStorage.getItem("token")||sessionStorage.getItem("token");if(!l){s.setIsAuthenticated(!1);return}if(typeof l!="string"||l.length<10){localStorage.removeItem("token"),sessionStorage.removeItem("token"),s.setIsAuthenticated(!1);return}const c=Le.replace("www.",""),i=yield fetch(`${c}auth/verify`,{method:"GET",headers:{Authorization:`Bearer ${l}`,Accept:"application/json"},mode:"cors",credentials:"include"});if(i.status===429)throw new Error("Zbyt wiele prób weryfikacji. Spróbuj ponownie za chwilę.");let a,w="";try{if(w=yield i.text(),!w)throw new Error("Pusta odpowiedź serwera");a=JSON.parse(w)}catch(h){throw new Error("Nieprawidłowa odpowiedź serwera: "+h)}if((a==null?void 0:a.status)==="success"&&(a!=null&&a.data)){s.setUser(a.data.user||a.data),s.setIsAuthenticated(!0);return}if(!i.ok)throw new Error((a==null?void 0:a.error)||(a==null?void 0:a.message)||"Nieznany błąd weryfikacji tokenu");if(!(a!=null&&a.user))throw new Error("Odpowiedź API nie zawiera danych użytkownika");s.setUser(a.user),s.setIsAuthenticated(!0)}catch(l){s.setError(l instanceof Error?l.message:"Wystąpił błąd podczas weryfikacji tokenu"),s.setIsAuthenticated(!1),localStorage.removeItem("token"),sessionStorage.removeItem("token")}finally{o(!1)}});g.useEffect(()=>{u()},[]);const n=()=>{localStorage.removeItem("token"),sessionStorage.removeItem("token"),s.setIsAuthenticated(!1),s.setUser(null),r("/")},m=z(x({},s),{isAuthChecking:d,token:localStorage.getItem("token")||sessionStorage.getItem("token")});return z(x(x({},m),t),{logout:n})},Ue=s=>{const t=g.useRef(null),r=g.useRef({x:0,y:0}),d=g.useRef(),o=g.useMemo(()=>100,[]),u=g.useMemo(()=>[new C("#f7df1e",2),new C("#f7df1e",2),new C("#f7df1e",2)],[]),n=g.useCallback(()=>{const c=new be(.5,0),i=new ke({color:new V("#f7df1e").multiplyScalar(.5),emissive:new V("#f7df1e").multiplyScalar(.3),emissiveIntensity:.5,transparent:!0,opacity:.8,shininess:90});return new Se(c,i)},[]),m=g.useCallback(c=>{var a;const i=(a=s.current)==null?void 0:a.getBoundingClientRect();i&&(r.current={x:(c.clientX-i.left)/i.width*2-1,y:-((c.clientY-i.top)/i.height)*2+1})},[]),l=g.useCallback((c,i)=>{if(!s.current)return;const{width:a,height:w}=s.current.getBoundingClientRect();c.setSize(a,w),i.aspect=a/w,i.updateProjectionMatrix()},[]);return g.useEffect(()=>{if(!t.current||!s.current)return;const c=new ve,i=new ze(75,1,.1,1e3),a=new Ne({canvas:t.current,alpha:!0,antialias:!0,powerPreference:"high-performance"});a.setClearColor(0,0);const w=new Ee,h=[];for(let y=0;y<o;y++){const b=n();b.position.set((Math.random()-.5)*50,(Math.random()-.5)*50,(Math.random()-.5)*50),b.rotation.set(Math.random()*Math.PI,Math.random()*Math.PI,Math.random()*Math.PI),h.push(b),w.add(b)}c.add(w),u.forEach((y,b)=>{y.position.set(Math.cos(b*Math.PI*2/3)*30,Math.sin(b*Math.PI*2/3)*30,15),c.add(y)}),c.add(new Pe("#ffffff",.5)),i.position.z=30;const p=()=>{d.current=requestAnimationFrame(p),h.forEach((y,b)=>{const te=Date.now()*.001+b;y.position.y+=Math.sin(te*.5)*.02,y.rotation.x+=.01,y.rotation.y+=.01,y.rotation.z+=.01}),w.rotation.y+=.001,w.rotation.x=r.current.y*.2,w.rotation.z=r.current.x*.2,a.render(c,i)},S=()=>l(a,i);return window.addEventListener("mousemove",m),window.addEventListener("resize",S),S(),p(),()=>{d.current&&cancelAnimationFrame(d.current),window.removeEventListener("mousemove",m),window.removeEventListener("resize",S),a.dispose(),c.clear()}},[o,u,n,m,l]),t},Ge=()=>{const s=g.useRef(null),t=Ue(s);return e.jsxs(e.Fragment,{children:[e.jsx(k.div,{ref:s,initial:{opacity:0},animate:{opacity:.3},className:"absolute inset-0 pointer-events-none",children:e.jsx("canvas",{ref:t,className:"w-full h-full"})}),e.jsxs("div",{className:"absolute top-0 left-0 w-full h-full overflow-hidden pointer-events-none",children:[e.jsx("div",{className:"absolute -top-20 -left-20 w-96 h-96 bg-[#f7df1e]/10 rounded-full blur-3xl"}),e.jsx("div",{className:"absolute -bottom-40 -right-20 w-[30rem] h-[30rem] bg-[#f7df1e]/5 rounded-full blur-3xl"}),e.jsx("div",{className:"absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-[40rem] h-[40rem] bg-[#1a1a1a]/50 rounded-full blur-3xl"}),e.jsx("div",{className:"absolute inset-0 bg-[url('/grid-pattern.svg')] bg-center opacity-5"}),e.jsx("div",{className:"absolute inset-0 bg-gradient-to-b from-[#1a1a1a]/0 via-[#1a1a1a]/0 to-[#1a1a1a]/80"})]})]})},Oe=()=>e.jsx("div",{className:"hidden md:block",children:e.jsx(fe,{position:"left",title:"CodeLinesJS App",className:"transform scale-90"})}),A=({text:s})=>e.jsx("div",{className:"flex items-center justify-center md:justify-start bg-[#f7df1e]/10 rounded-lg p-3",children:e.jsxs("span",{className:"text-[#f7df1e] font-medium",children:["✓ ",s]})}),Ve=()=>e.jsxs("div",{className:"grid grid-cols-1 sm:grid-cols-2 gap-4",children:[e.jsx(A,{text:"Interaktywne lekcje"}),e.jsx(A,{text:"Wyzwania i gry"}),e.jsx(A,{text:"Śledzenie postępów"}),e.jsx(A,{text:"Społeczność"})]}),Ze=()=>e.jsxs("div",{className:"text-center md:text-left max-w-xl",children:[e.jsxs("h1",{className:"text-3xl md:text-4xl lg:text-5xl font-bold text-white mb-6",children:["Odkryj ",e.jsx("span",{className:"text-[#f7df1e]",children:"JavaScript"})," w nowy sposób"]}),e.jsx("p",{className:"text-lg md:text-xl text-gray-300 mb-8",children:"Ucz się, programuj i baw się jednocześnie. Nasza platforma łączy teorię z praktyką, oferując interaktywne lekcje, wyzwania i gry."}),e.jsx(Ve,{})]}),$e=()=>e.jsxs(k.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.5,delay:.2},className:"w-full md:w-3/5 flex flex-col md:flex-row items-center gap-6 md:gap-12 pr-0 md:pr-8",children:[e.jsx(Oe,{}),e.jsx(Ze,{})]}),Je=({activeTab:s,onTabChange:t})=>e.jsxs("div",{className:"flex items-center justify-center gap-4 mb-8",children:[e.jsx(_,{isActive:s==="login",onClick:()=>t("login"),icon:e.jsx(le,{}),label:"Logowanie"}),e.jsx(_,{isActive:s==="register",onClick:()=>t("register"),icon:e.jsx(ce,{}),label:"Rejestracja"}),e.jsx(_,{isActive:s==="forgot",onClick:()=>t("forgot"),icon:e.jsx(de,{}),label:"Reset hasła"})]}),_=({isActive:s,onClick:t,icon:r,label:d})=>e.jsxs(k.button,{whileHover:{scale:1.05},whileTap:{scale:.95},onClick:t,className:`flex flex-col items-center gap-2 p-3 rounded-lg transition-colors ${s?"bg-js/20 text-js":"text-gray-400 hover:text-js"}`,children:[r,e.jsx("span",{className:"text-sm font-medium",children:d})]}),De=g.lazy(()=>B(()=>Promise.resolve().then(()=>ds),void 0)),qe=g.lazy(()=>B(()=>Promise.resolve().then(()=>ws),void 0)),Ke=g.lazy(()=>B(()=>Promise.resolve().then(()=>xs),void 0)),Xe=()=>{const[s,t]=g.useState("login");return e.jsxs(k.div,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},transition:{duration:.5},className:"w-full max-w-md bg-dark/30 backdrop-blur-xl rounded-2xl p-8 border border-js/20 shadow-xl shadow-black/20",role:"main","aria-label":"Formularz autoryzacji",children:[e.jsxs("div",{className:"flex flex-col justify-center items-center mb-6",children:[e.jsx("span",{className:"block text-2xl bg-[#f7df1e] text-black px-2 py-1 rounded",children:"JS"}),e.jsx("h2",{className:"mt-2 text-xl font-bold text-js",children:"CodeLinesJS"})]}),e.jsxs("h1",{className:"text-2xl md:text-3xl font-bold text-center text-js mb-6",children:[s==="login"&&"Zaloguj się",s==="register"&&"Dołącz do nas",s==="forgot"&&"Odzyskaj hasło"]}),e.jsx(Je,{activeTab:s,onTabChange:t}),e.jsx(g.Suspense,{fallback:e.jsx("div",{className:"text-center py-4",children:"Ładowanie..."}),children:e.jsxs(Me,{mode:"wait",children:[s==="login"&&e.jsx(De,{},"login"),s==="register"&&e.jsx(qe,{},"register"),s==="forgot"&&e.jsx(Ke,{},"forgot")]})}),e.jsx("div",{className:"mt-6 text-center text-sm text-gray-400",children:e.jsxs("p",{children:["Logując się, akceptujesz"," ",e.jsx(E,{to:"/polityka-prywatnosci",className:"text-js hover:underline",children:"Politykę Prywatności"}),"."]})})]})},Ye=({children:s,title:t,subtitle:r})=>e.jsxs("div",{className:"w-full md:w-2/5 flex justify-center md:justify-end relative mt-12 md:mt-0",children:[e.jsx("div",{className:"absolute -z-10 top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-64 h-64 bg-[#f7df1e]/10 rounded-full blur-3xl"}),s?e.jsxs("div",{className:"w-full max-w-md bg-dark/30 backdrop-blur-xl rounded-2xl p-8 border border-js/20",children:[e.jsxs("div",{className:"flex flex-col justify-center items-center mb-8",children:[e.jsx("span",{className:"block text-2xl bg-[#f7df1e] text-black px-2 py-1 rounded",children:"JS"}),e.jsx("h2",{className:"mt-2 text-xl font-bold text-js",children:"CodeLinesJS"})]}),t&&e.jsx("h1",{className:"text-3xl font-bold text-center text-js mb-4",children:t}),r&&e.jsx("p",{className:"text-center text-gray-400 mb-8",children:r}),s]}):e.jsx(Xe,{})]}),$=({children:s,title:t,subtitle:r}={})=>e.jsxs("main",{className:"min-h-screen w-full relative bg-gradient-to-b from-[#1a1a1a] via-[#242424] to-[#1a1a1a] flex items-center justify-center py-16 overflow-hidden",children:[e.jsxs(me,{children:[e.jsx("title",{children:" Logowanie | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Logowanie do platformy CodeLinesJS - dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."})]}),e.jsx(je,{className:"relative z-10",children:e.jsxs("div",{className:"flex flex-col md:flex-row items-center justify-between gap-8 md:gap-20 w-full max-w-7xl mx-auto",children:[e.jsx($e,{}),e.jsx(ye,{children:e.jsx(Ye,{children:s,title:t,subtitle:r})})]})}),e.jsx(Ge,{})]}),vs=()=>{const[s,t]=g.useState(!1);return{isVisible:s,toggleVisibility:()=>t(d=>!d),inputType:s?"text":"password"}},Qe=({errorMessage:s})=>e.jsxs(k.div,{initial:{opacity:0},animate:{opacity:1},className:"space-y-6",children:[e.jsxs("div",{className:"p-4 bg-red-500/10 border border-red-500/20 rounded-lg text-red-400 text-sm flex items-start",children:[e.jsx(J,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:"Błąd tokenu resetowania hasła"}),e.jsx("p",{children:s||"Nieprawidłowy lub wygasły token resetowania hasła. Sprawdź, czy link jest poprawny lub spróbuj ponownie zresetować hasło."})]})]}),e.jsxs("div",{className:"text-center mt-6",children:[e.jsx("p",{className:"text-gray-400 mb-4",children:"Możesz spróbować ponownie zresetować hasło:"}),e.jsx(E,{to:"/logowanie",className:"text-js hover:underline",children:"Wróć do strony logowania"})]})]}),es=({message:s})=>s?e.jsxs(e.Fragment,{children:[e.jsx(q,{message:s,title:"Hasło zostało zmienione"}),e.jsxs("div",{className:"text-center mt-4",children:[e.jsx("p",{className:"text-gray-400 mb-2",children:"Za chwilę zostaniesz przekierowany do strony logowania."}),e.jsx(E,{to:"/logowanie",className:"text-js hover:underline",children:"Przejdź do logowania"})]})]}):null,ss=f.object({email:f.string().min(1,"Email jest wymagany").email("Nieprawidłowy format email"),password:f.string().min(1,"Hasło jest wymagane"),rememberMe:f.boolean().optional()}),ts=f.object({username:f.string().min(3,"Nazwa użytkownika musi mieć minimum 3 znaki").max(20,"Nazwa użytkownika nie może przekraczać 20 znaków").regex(/^[a-zA-Z0-9_]+$/,"Nazwa użytkownika może zawierać tylko litery, cyfry i podkreślenia"),email:f.string().email("Nieprawidłowy format adresu email"),password:f.string().min(8,"Hasło musi mieć minimum 8 znaków").regex(/[A-Z]/,"Hasło musi zawierać przynajmniej jedną wielką literę").regex(/[a-z]/,"Hasło musi zawierać przynajmniej jedną małą literę").regex(/[0-9]/,"Hasło musi zawierać przynajmniej jedną cyfrę"),confirmPassword:f.string(),acceptPrivacy:f.boolean().refine(s=>s===!0,{message:"Musisz zaakceptować politykę prywatności"})}).refine(s=>s.password===s.confirmPassword,{message:"Hasła muszą być identyczne",path:["confirmPassword"]}),rs=f.object({email:f.string().min(1,"Email jest wymagany").email("Nieprawidłowy format email")}),as=f.object({password:f.string().min(8,"Hasło musi mieć minimum 8 znaków").regex(/[A-Z]/,"Hasło musi zawierać przynajmniej jedną wielką literę").regex(/[a-z]/,"Hasło musi zawierać przynajmniej jedną małą literę").regex(/[0-9]/,"Hasło musi zawierać przynajmniej jedną cyfrę"),confirmPassword:f.string()}).refine(s=>s.password===s.confirmPassword,{message:"Hasła muszą być identyczne",path:["confirmPassword"]}),H=({initialError:s=null,resetOnSuccess:t=!0,showToasts:r=!0}={})=>{const[d,o]=g.useState(null),[u,n]=g.useState(s);g.useEffect(()=>{s!=null&&n(s)},[s]);const m=g.useCallback(a=>{o(a),n(null),r&&O.success(a),t&&setTimeout(()=>{o(null)},5e3)},[t,r]),l=g.useCallback(a=>{n(a),o(null),r&&O.error(a)},[r]),c=g.useCallback(a=>{var h;let w="";if(a instanceof Error)w=a.message;else if(typeof a=="string")w=a;else if(a&&typeof a=="object"){const p=a;w=p.message||p.error||p.errors&&((h=p.errors[0])==null?void 0:h.message)||"Wystąpił nieznany błąd. Spróbuj ponownie później."}else w="Wystąpił nieznany błąd. Spróbuj ponownie później.";return l(w),w},[l]),i=g.useCallback(()=>{o(null),n(null)},[]);return{successMessage:d,errorMessage:u,setSuccess:m,setError:l,handleError:c,resetStatus:i,hasSuccess:!!d,hasError:!!u}},os=()=>{const{token:s}=D(),{resetPassword:t,loading:r,error:d}=M(),o=H({initialError:d}),[u,n]=g.useState(!1),{register:m,handleSubmit:l,formState:{errors:c},reset:i}=L({resolver:F(as),mode:"onSubmit"});return g.useEffect(()=>{s?s.length<10&&(n(!0),o.setError("Token resetowania hasła jest nieprawidłowy. Sprawdź, czy link jest poprawny.")):(n(!0),o.setError("Brak tokenu resetowania hasła. Sprawdź, czy link jest poprawny."))},[s,o]),x({register:m,errors:c,loading:r,tokenError:u,handleSubmit:l(w=>j(void 0,null,function*(){if(!s){o.setError("Brak tokenu resetowania hasła. Sprawdź, czy link jest poprawny.");return}try{o.resetStatus();const h=yield t(s,w.password,w.confirmPassword);o.setSuccess(h),i()}catch(h){const p=o.handleError(h);typeof p=="string"&&(p.includes("token")||p.includes("Token")||p.includes("wygasł"))&&n(!0)}}))},o)},ns=()=>{var m,l;const{register:s,errors:t,loading:r,handleSubmit:d,errorMessage:o,successMessage:u,tokenError:n}=os();return n?e.jsx(Qe,{errorMessage:o}):e.jsxs(T,{onSubmit:d,children:[o&&e.jsx(K,{message:o,title:"Błąd resetowania hasła"}),e.jsx(es,{message:u}),!u&&e.jsxs(e.Fragment,{children:[e.jsx("p",{className:"text-gray-400 text-sm mb-6",children:"Wprowadź nowe hasło dla swojego konta. Hasło musi zawierać co najmniej 8 znaków, w tym jedną wielką literę, jedną małą literę i jedną cyfrę."}),e.jsx(I,x({label:"Nowe hasło",error:(m=t.password)==null?void 0:m.message},s("password"))),e.jsx(I,x({label:"Potwierdź nowe hasło",error:(l=t.confirmPassword)==null?void 0:l.message},s("confirmPassword"))),e.jsx(X,{type:"submit",className:"w-full",disabled:r,children:r?"Resetowanie hasła...":"Zresetuj hasło"})]})]})},is=()=>e.jsxs(k.div,{initial:{opacity:0},animate:{opacity:1},className:"space-y-6",children:[e.jsxs("div",{className:"p-4 bg-red-500/10 border border-red-500/20 rounded-lg text-red-400 text-sm flex items-start",children:[e.jsx(J,{className:"mr-2 mt-0.5 flex-shrink-0"}),e.jsxs("div",{children:[e.jsx("p",{className:"font-semibold mb-1",children:"Błąd tokenu resetowania hasła"}),e.jsx("p",{children:"Nieprawidłowy lub wygasły token resetowania hasła. Sprawdź, czy link jest poprawny lub spróbuj ponownie zresetować hasło."})]})]}),e.jsxs("div",{className:"text-center mt-6",children:[e.jsx("p",{className:"text-gray-400 mb-4",children:"Możesz spróbować ponownie zresetować hasło na stronie logowania:"}),e.jsxs(E,{to:"/logowanie",className:"inline-flex items-center text-js hover:underline",children:[e.jsx(ue,{className:"mr-2"})," Wróć do strony logowania"]})]})]}),ls=()=>{const{token:s}=D();return s?e.jsx($,{title:"Resetowanie hasła",subtitle:"Wprowadź nowe hasło dla swojego konta",children:e.jsx(ns,{})}):e.jsx($,{title:"Nieprawidłowy link",subtitle:"Link do resetowania hasła jest nieprawidłowy lub wygasł.",children:e.jsx(is,{})})},zs=Object.freeze(Object.defineProperty({__proto__:null,ResetPasswordPage:ls},Symbol.toStringTag,{value:"Module"})),Q=g.memo(({rememberMe:s})=>{const{loginWithGoogle:t}=M(),[r,d]=g.useState(null),o=n=>{t(n,s)},u=()=>{d("Logowanie przez Google nie powiodło się. Spróbuj ponownie później.")};return e.jsxs(e.Fragment,{children:[r&&e.jsx("div",{className:"mb-4 p-3 rounded-lg bg-red-500/10 border border-red-500/20 text-red-400 text-sm",children:r}),e.jsx("div",{className:"flex flex-col items-center",children:e.jsx(ge,{onSuccess:o,onError:u,theme:"filled_black",shape:"pill",text:"continue_with",locale:"pl",useOneTap:!1,context:"signin",type:"standard",size:"large",width:"100%"})})]})});Q.displayName="GoogleLoginButton";const ee=g.memo(({register:s})=>e.jsx("div",{className:"flex items-center",children:e.jsxs("label",{className:"flex items-center gap-2 cursor-pointer group",children:[e.jsxs("div",{className:"relative",children:[e.jsx("input",x({type:"checkbox",className:"peer sr-only"},s("rememberMe"))),e.jsx("div",{className:"w-4 h-4 border border-js/30 rounded bg-dark/70 peer-checked:bg-js peer-checked:border-js transition-colors"}),e.jsx("div",{className:"absolute inset-0 flex items-center justify-center text-black scale-0 peer-checked:scale-100 transition-transform",children:e.jsx("svg",{xmlns:"http://www.w3.org/2000/svg",className:"h-3 w-3",viewBox:"0 0 20 20",fill:"currentColor",children:e.jsx("path",{fillRule:"evenodd",d:"M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z",clipRule:"evenodd"})})})]}),e.jsx("span",{className:"text-sm text-gray-400 group-hover:text-gray-300 transition-colors",children:"Zapamiętaj mnie"})]})}));ee.displayName="RememberMeCheckbox";const cs=()=>{var a,w;const{login:s,loading:t,error:r}=M(),{register:d,handleSubmit:o,formState:{errors:u},watch:n}=L({resolver:F(ss),mode:"onSubmit"}),m=n("rememberMe",!1)||!1,[l,c]=g.useState(!1),i=h=>j(void 0,null,function*(){yield s(h.email,h.password,h.rememberMe)});return e.jsxs(k.form,{initial:{opacity:0,y:20},animate:{opacity:1,y:0},onSubmit:o(i),className:"space-y-5",children:[r&&e.jsx(R,{message:r,withIcon:!0,variant:"standard"}),e.jsx(N,x({type:"email",label:"Email",placeholder:"twoj@email.com",icon:e.jsx(W,{}),error:(a=u.email)==null?void 0:a.message},d("email"))),e.jsx(N,z(x({type:l?"text":"password",label:"Hasło",placeholder:"••••••••",icon:e.jsx(we,{}),error:(w=u.password)==null?void 0:w.message},d("password")),{rightIcon:e.jsx("button",{type:"button",onClick:()=>c(!l),className:"focus:outline-none text-gray-400 hover:text-js transition-colors",children:l?e.jsx(he,{}):e.jsx(pe,{})})})),e.jsx("div",{className:"flex items-center justify-between",children:e.jsx(ee,{register:d})}),e.jsx("div",{className:"pt-2",children:e.jsx(Y,{isLoading:t,loadingText:"Logowanie...",children:"Zaloguj się"})}),e.jsxs("div",{className:"relative flex items-center justify-center my-4",children:[e.jsx("div",{className:"border-t border-gray-600 w-full"}),e.jsx("span",{className:"bg-dark/50 px-3 text-sm text-gray-400 absolute",children:"lub"})]}),e.jsx(Q,{rememberMe:m})]})},ds=Object.freeze(Object.defineProperty({__proto__:null,default:cs},Symbol.toStringTag,{value:"Module"})),se=g.memo(({register:s,error:t})=>e.jsxs("div",{className:"space-y-1",children:[e.jsxs("label",{className:"flex items-start gap-2 cursor-pointer group",children:[e.jsx("input",x({type:"checkbox",className:`
-          mt-1 w-4 h-4 rounded bg-dark text-js 
-          transition-colors duration-200 ease-in-out
-          ${t?"border-red-500 focus:ring-red-500/20":"border-js/20 focus:ring-js/20 hover:border-js/30"}
-        `},s)),e.jsxs("span",{className:"text-sm text-gray-400 group-hover:text-gray-300 transition-colors",children:["Akceptuję"," ",e.jsx(E,{to:"/polityka-prywatnosci",className:"text-js hover:text-js/80 underline transition-colors",target:"_blank",rel:"noopener noreferrer",children:"politykę prywatności"})]})]}),t&&e.jsx(k.p,{initial:{opacity:0,y:-10},animate:{opacity:1,y:0},className:"text-sm text-red-400",children:t})]}));se.displayName="PrivacyPolicyCheckbox";const ms=({schema:s,defaultValues:t,mode:r="onSubmit",resetOnSuccess:d=!0,onSuccess:o,onError:u})=>{const[n,m]=g.useState(!1),l=H({resetOnSuccess:d}),c=L({resolver:F(s),defaultValues:t,mode:r}),i=g.useCallback(p=>j(void 0,null,function*(){try{m(!0),l.resetStatus(),o&&(yield o(p))}catch(S){l.handleError(S),u&&u(S)}finally{m(!1)}}),[o,u,l]),a=g.useCallback(p=>{i(p)},[i]),w=g.useCallback(p=>{console.error("Form validation errors:",p),l.setError("Formularz zawiera błędy. Sprawdź wprowadzone dane.")},[l]),h=c.handleSubmit(a,w);return z(x({},c),{isSubmitting:n,errorMessage:l.errorMessage,successMessage:l.successMessage,onFormSubmit:h,handleSubmitWithValidation:i})},us=()=>{const{register:s,loading:t,error:r}=M(),[d,o]=g.useState(!1),[u,n]=g.useState(null),m=ms({schema:ts,mode:"onChange",defaultValues:{acceptPrivacy:!1},onSuccess(c){return j(this,null,function*(){try{o(!0),n(null),yield s(c.email,c.password,c.username)}catch(i){const a=i instanceof Error?i.message:"Wystąpił błąd podczas rejestracji";n(a)}finally{o(!1)}})}});return g.useEffect(()=>{r&&n(r)},[r]),{register:m.register,errors:m.formState.errors,loading:t||d||m.isSubmitting,handleSubmit:m.onFormSubmit,errorMessage:u||r||m.errorMessage,successMessage:m.successMessage}},gs=()=>{var u,n,m,l,c;const{register:s,errors:t,loading:r,handleSubmit:d,errorMessage:o}=us();return e.jsxs(T,{onSubmit:d,children:[o&&e.jsx(R,{message:o,variant:"alert",title:"Błąd rejestracji",className:"mb-6 text-base border-2"}),e.jsx(N,x({type:"text",label:"Nazwa użytkownika",placeholder:"jankowalski",icon:e.jsx(xe,{}),error:(u=t.username)==null?void 0:u.message},s("username"))),e.jsx(N,x({type:"email",label:"Email",placeholder:"twoj@email.com",icon:e.jsx(W,{}),error:(n=t.email)==null?void 0:n.message},s("email"))),e.jsx(I,x({label:"Hasło",error:(m=t.password)==null?void 0:m.message},s("password"))),e.jsx(I,x({label:"Potwierdź hasło",error:(l=t.confirmPassword)==null?void 0:l.message},s("confirmPassword"))),e.jsx(se,{register:s("acceptPrivacy"),error:(c=t.acceptPrivacy)==null?void 0:c.message}),e.jsx(Y,{isLoading:r,loadingText:"Rejestracja...",children:"Zarejestruj się"}),o&&e.jsx("div",{className:"mt-4",children:e.jsx(R,{message:o,withIcon:!0,className:"text-sm"})})]})},ws=Object.freeze(Object.defineProperty({__proto__:null,default:gs},Symbol.toStringTag,{value:"Module"})),hs=()=>{const{forgotPassword:s,loading:t,error:r}=M(),d=H({initialError:r}),{register:o,handleSubmit:u,reset:n,formState:{errors:m}}=L({resolver:F(rs)});return x({register:o,errors:m,loading:t,handleSubmit:u(c=>j(void 0,null,function*(){try{d.resetStatus();const i=yield s(c.email);d.setSuccess(i),n()}catch(i){d.handleError(i)}}))},d)},ps=()=>{var n;const{register:s,errors:t,loading:r,handleSubmit:d,errorMessage:o,successMessage:u}=hs();return e.jsxs(T,{onSubmit:d,children:[o&&e.jsx(K,{message:o,title:"Błąd resetowania hasła"}),u&&e.jsx(q,{message:u}),!u&&e.jsxs(e.Fragment,{children:[e.jsx("p",{className:"text-gray-400 text-sm mb-6",children:"Podaj swój adres email, a wyślemy Ci link do zresetowania hasła."}),e.jsx(N,x({type:"email",label:"Email",placeholder:"twoj@email.com",icon:e.jsx(W,{}),error:(n=t.email)==null?void 0:n.message},s("email"))),e.jsx(X,{type:"submit",className:"w-full",disabled:r,children:r?"Wysyłanie...":"Wyślij link resetujący"})]})]})},xs=Object.freeze(Object.defineProperty({__proto__:null,default:ps},Symbol.toStringTag,{value:"Module"}));export{Le as A,B as _,vs as a,$ as b,zs as i,M as u};
