+++ server/public/assets/app-dashboard-BQ6AEXi0.js
@@ -1,5 +1,5 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/app-components-DHLBCYYN.js","assets/vendor-react-RXQ05gJ2.js","assets/vendor-other-C1_ffh8T.js","assets/vendor-monaco-Dk8M9SDd.js","assets/vendor-framer-motion-SZOJYst1.js","assets/app-auth-DYLPzh_K.js","assets/app-services-tJo8l8LQ.js"])))=>i.map(i=>d[i]);
-var Ei=Object.defineProperty,Ci=Object.defineProperties;var Ai=Object.getOwnPropertyDescriptors;var ds=Object.getOwnPropertySymbols;var $i=Object.prototype.hasOwnProperty,Ri=Object.prototype.propertyIsEnumerable;var ms=(s,t,a)=>t in s?Ei(s,t,{enumerable:!0,configurable:!0,writable:!0,value:a}):s[t]=a,k=(s,t)=>{for(var a in t||(t={}))$i.call(t,a)&&ms(s,a,t[a]);if(ds)for(var a of ds(t))Ri.call(t,a)&&ms(s,a,t[a]);return s},z=(s,t)=>Ci(s,Ai(t));var L=(s,t,a)=>new Promise((i,n)=>{var o=d=>{try{l(a.next(d))}catch(u){n(u)}},c=d=>{try{l(a.throw(d))}catch(u){n(u)}},l=d=>d.done?i(d.value):Promise.resolve(d.value).then(o,c);l((a=a.apply(s,t)).next())});import{_ as K,u as F,A}from"./app-auth-DYLPzh_K.js";import{r,R as Ce,j as e,S as xs,H as Y,T as Fi,U as us,y as O,s as _i,V as Di,q as Ke,x as Ze,C as ys,W as gs,X as fs,u as ce,Y as bs,E as G,Z as de,_ as ae,$ as Ve,t as Oi,B as ee,L as fe,a0 as Ii,a1 as ws,P as D,a2 as Ns,a3 as vs,a4 as V,c as X,a5 as ks,n as me,a6 as I,h as Z,a7 as Ss,a8 as Xe,a9 as Ye,aa as Bi,ab as Mi,ac as zs,ad as be,ae as P,af as Ls,ag as Gi,ah as Hi,ai as Ui,o as Ts,d as es,N as Ps,m as Es,aj as ss,e as Ae,g as Wi,ak as Ji,al as Cs,am as qi,an as Qi,f as Ki,ao as xe,ap as As,aq as $s,ar as Zi,as as Vi,at as Xi,au as Yi,av as er,aw as sr,ax as tr,ay as ts,az as as,aA as ar,aB as ir,aC as rr,aD as nr,aE as or,aF as lr,aG as cr,aH as dr,aI as mr}from"./vendor-react-RXQ05gJ2.js";import{L as Ue,u as Rs,M as xr,e as is,f as U,g as ur,h as pr,c as le,B as ye}from"./app-components-DHLBCYYN.js";import{m,A as B,u as hr}from"./vendor-framer-motion-SZOJYst1.js";import{f as Fs,p as _s,z as _,t as $e,c as We}from"./vendor-other-C1_ffh8T.js";import{R as jr,L as yr,C as gr,X as fr,Y as br,T as wr,a as Nr}from"./vendor-charts-Czvt6lRK.js";import{l as vr}from"./vendor-monaco-Dk8M9SDd.js";import{h as kr}from"./app-services-tJo8l8LQ.js";const Sr={activeView:"community"},zr=(s,t)=>{switch(t.type){case"SET_ACTIVE_VIEW":return z(k({},s),{activeView:t.payload});default:return s}},Lr=()=>{const[s,t]=r.useReducer(zr,Sr),a=r.useCallback(i=>{t({type:"SET_ACTIVE_VIEW",payload:i})},[]);return{state:s,setActiveView:a}},Ds=r.memo(()=>{const t=Ce().pathname,a=t==="/dashboard/community"||t==="/dashboard/community/"||/^\/dashboard\/community\/?$/.test(t),i=t.includes("/dashboard/community/ranking");return e.jsxs("div",{className:"flex gap-4 mb-8 pb-2",children:[e.jsxs(xs,{to:"/dashboard/community",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${a?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-newspaper"}),"Aktualności"]}),e.jsxs(xs,{to:"/dashboard/community/ranking",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${i?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-trophy"}),"Ranking"]})]})});Ds.displayName="CommunityNavigation";const Tr=r.lazy(()=>K(()=>Promise.resolve().then(()=>bo),void 0)),Pr=r.lazy(()=>K(()=>Promise.resolve().then(()=>vo),void 0)),Os=r.memo(()=>{const s=Ce(),{state:{activeView:t},setActiveView:a}=Lr();return r.useEffect(()=>{const i=s.pathname.split("community").pop()||"community";i!==t&&a(i)},[s.pathname,t,a]),e.jsxs("div",{className:"p-8 w-full min-h-screen bg-dark/50 backdrop-blur-sm",children:[e.jsxs(Y,{children:[e.jsx("title",{children:"Community | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Community CodeLinesJS - dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."})]}),e.jsx("div",{className:"max-w-7xl mx-auto",children:e.jsxs("div",{className:"bg-dark/30 backdrop-blur-sm rounded-xl border border-js/10 p-6 shadow-lg",children:[e.jsx(Ds,{}),e.jsxs(Fi,{children:[e.jsx(us,{index:!0,element:e.jsx(r.Suspense,{fallback:e.jsx(Ue,{text:"Ładowanie aktualności..."}),children:e.jsx(Tr,{})})}),e.jsx(us,{path:"ranking",element:e.jsx(r.Suspense,{fallback:e.jsx(Ue,{text:"Ładowanie rankingu..."}),children:e.jsx(Pr,{})})})]})]})})]})});Os.displayName="CommunitySection";const ll=Object.freeze(Object.defineProperty({__proto__:null,CommunitySection:Os},Symbol.toStringTag,{value:"Module"})),Er=()=>{const[s,t]=r.useState(!1),[a,i]=r.useState("dashboard"),n=Ce();return r.useEffect(()=>{const c=n.pathname.split("/").pop()||"dashboard";i(c)},[n]),r.useEffect(()=>{const o=()=>{window.innerWidth<768&&t(!1)};return window.addEventListener("resize",o),o(),()=>window.removeEventListener("resize",o)},[]),{isExpanded:s,setIsExpanded:t,activeItem:a,setActiveItem:i}},Cr={expanded:{width:"240px",transition:{type:"spring",stiffness:300,damping:30}},collapsed:{width:"100px",transition:{type:"spring",stiffness:300,damping:30}}},rs={expanded:{opacity:1,x:0,transition:{type:"spring",stiffness:300,damping:30}},collapsed:{opacity:0,x:-10,transition:{duration:.2}}},Is=r.memo(({isExpanded:s})=>e.jsx("div",{className:"px-4 mb-8",children:e.jsxs(m.div,{initial:{opacity:0},animate:{opacity:1},className:"flex items-center gap-3",children:[e.jsx("div",{className:"w-10 h-10 min-w-[40px] rounded-xl bg-js/20 flex items-center justify-center",children:e.jsx(O,{className:"text-xl text-js"})}),e.jsx(B,{mode:"wait",children:s&&e.jsx(m.span,{variants:rs,initial:"collapsed",animate:"expanded",exit:"collapsed",className:"font-bold text-js font-space",children:"CodeLinesJS"})})]})}));Is.displayName="NavigationLogo";const ns=r.memo(({icon:s,label:t,isActive:a=!1,isExpanded:i,onClick:n,variant:o="default",id:c})=>e.jsxs(m.button,{onClick:n,"data-testid":`nav-button-${t==null?void 0:t.toLowerCase().replace(/\s/g,"-")}`,className:`
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/app-components-DiGU-x_9.js","assets/vendor-react-RXQ05gJ2.js","assets/vendor-other-C1_ffh8T.js","assets/vendor-monaco-Dk8M9SDd.js","assets/vendor-framer-motion-SZOJYst1.js","assets/app-auth-CQDHbRjF.js","assets/app-services-BaprE51I.js"])))=>i.map(i=>d[i]);
+var Ei=Object.defineProperty,Ci=Object.defineProperties;var Ai=Object.getOwnPropertyDescriptors;var ds=Object.getOwnPropertySymbols;var $i=Object.prototype.hasOwnProperty,Ri=Object.prototype.propertyIsEnumerable;var ms=(s,t,a)=>t in s?Ei(s,t,{enumerable:!0,configurable:!0,writable:!0,value:a}):s[t]=a,k=(s,t)=>{for(var a in t||(t={}))$i.call(t,a)&&ms(s,a,t[a]);if(ds)for(var a of ds(t))Ri.call(t,a)&&ms(s,a,t[a]);return s},z=(s,t)=>Ci(s,Ai(t));var L=(s,t,a)=>new Promise((i,n)=>{var o=d=>{try{l(a.next(d))}catch(u){n(u)}},c=d=>{try{l(a.throw(d))}catch(u){n(u)}},l=d=>d.done?i(d.value):Promise.resolve(d.value).then(o,c);l((a=a.apply(s,t)).next())});import{_ as K,u as F,A}from"./app-auth-CQDHbRjF.js";import{r,R as Ce,j as e,S as xs,H as Y,T as Fi,U as us,y as O,s as _i,V as Di,q as Ke,x as Ze,C as ys,W as gs,X as fs,u as ce,Y as bs,E as G,Z as de,_ as ae,$ as Ve,t as Oi,B as ee,L as fe,a0 as Ii,a1 as ws,P as D,a2 as Ns,a3 as vs,a4 as V,c as X,a5 as ks,n as me,a6 as I,h as Z,a7 as Ss,a8 as Xe,a9 as Ye,aa as Bi,ab as Mi,ac as zs,ad as be,ae as P,af as Ls,ag as Gi,ah as Hi,ai as Ui,o as Ts,d as es,N as Ps,m as Es,aj as ss,e as Ae,g as Wi,ak as Ji,al as Cs,am as qi,an as Qi,f as Ki,ao as xe,ap as As,aq as $s,ar as Zi,as as Vi,at as Xi,au as Yi,av as er,aw as sr,ax as tr,ay as ts,az as as,aA as ar,aB as ir,aC as rr,aD as nr,aE as or,aF as lr,aG as cr,aH as dr,aI as mr}from"./vendor-react-RXQ05gJ2.js";import{L as Ue,u as Rs,M as xr,e as is,f as U,g as ur,h as pr,c as le,B as ye}from"./app-components-DiGU-x_9.js";import{m,A as B,u as hr}from"./vendor-framer-motion-SZOJYst1.js";import{f as Fs,p as _s,z as _,t as $e,c as We}from"./vendor-other-C1_ffh8T.js";import{R as jr,L as yr,C as gr,X as fr,Y as br,T as wr,a as Nr}from"./vendor-charts-Czvt6lRK.js";import{l as vr}from"./vendor-monaco-Dk8M9SDd.js";import{h as kr}from"./app-services-BaprE51I.js";const Sr={activeView:"community"},zr=(s,t)=>{switch(t.type){case"SET_ACTIVE_VIEW":return z(k({},s),{activeView:t.payload});default:return s}},Lr=()=>{const[s,t]=r.useReducer(zr,Sr),a=r.useCallback(i=>{t({type:"SET_ACTIVE_VIEW",payload:i})},[]);return{state:s,setActiveView:a}},Ds=r.memo(()=>{const t=Ce().pathname,a=t==="/dashboard/community"||t==="/dashboard/community/"||/^\/dashboard\/community\/?$/.test(t),i=t.includes("/dashboard/community/ranking");return e.jsxs("div",{className:"flex gap-4 mb-8 pb-2",children:[e.jsxs(xs,{to:"/dashboard/community",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${a?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-newspaper"}),"Aktualności"]}),e.jsxs(xs,{to:"/dashboard/community/ranking",className:`flex items-center gap-2 px-6 py-3 font-semibold rounded-lg transition-all ${i?"bg-js/10 text-js border-b-2 border-js shadow-md":"text-gray-400 hover:text-js hover:bg-js/5"}`,children:[e.jsx("i",{className:"fas fa-trophy"}),"Ranking"]})]})});Ds.displayName="CommunityNavigation";const Tr=r.lazy(()=>K(()=>Promise.resolve().then(()=>bo),void 0)),Pr=r.lazy(()=>K(()=>Promise.resolve().then(()=>vo),void 0)),Os=r.memo(()=>{const s=Ce(),{state:{activeView:t},setActiveView:a}=Lr();return r.useEffect(()=>{const i=s.pathname.split("community").pop()||"community";i!==t&&a(i)},[s.pathname,t,a]),e.jsxs("div",{className:"p-8 w-full min-h-screen bg-dark/50 backdrop-blur-sm",children:[e.jsxs(Y,{children:[e.jsx("title",{children:"Community | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Community CodeLinesJS - dołącz do nas i rozwijaj swoje umiejętności w przyjaznym środowisku."})]}),e.jsx("div",{className:"max-w-7xl mx-auto",children:e.jsxs("div",{className:"bg-dark/30 backdrop-blur-sm rounded-xl border border-js/10 p-6 shadow-lg",children:[e.jsx(Ds,{}),e.jsxs(Fi,{children:[e.jsx(us,{index:!0,element:e.jsx(r.Suspense,{fallback:e.jsx(Ue,{text:"Ładowanie aktualności..."}),children:e.jsx(Tr,{})})}),e.jsx(us,{path:"ranking",element:e.jsx(r.Suspense,{fallback:e.jsx(Ue,{text:"Ładowanie rankingu..."}),children:e.jsx(Pr,{})})})]})]})})]})});Os.displayName="CommunitySection";const ll=Object.freeze(Object.defineProperty({__proto__:null,CommunitySection:Os},Symbol.toStringTag,{value:"Module"})),Er=()=>{const[s,t]=r.useState(!1),[a,i]=r.useState("dashboard"),n=Ce();return r.useEffect(()=>{const c=n.pathname.split("/").pop()||"dashboard";i(c)},[n]),r.useEffect(()=>{const o=()=>{window.innerWidth<768&&t(!1)};return window.addEventListener("resize",o),o(),()=>window.removeEventListener("resize",o)},[]),{isExpanded:s,setIsExpanded:t,activeItem:a,setActiveItem:i}},Cr={expanded:{width:"240px",transition:{type:"spring",stiffness:300,damping:30}},collapsed:{width:"100px",transition:{type:"spring",stiffness:300,damping:30}}},rs={expanded:{opacity:1,x:0,transition:{type:"spring",stiffness:300,damping:30}},collapsed:{opacity:0,x:-10,transition:{duration:.2}}},Is=r.memo(({isExpanded:s})=>e.jsx("div",{className:"px-4 mb-8",children:e.jsxs(m.div,{initial:{opacity:0},animate:{opacity:1},className:"flex items-center gap-3",children:[e.jsx("div",{className:"w-10 h-10 min-w-[40px] rounded-xl bg-js/20 flex items-center justify-center",children:e.jsx(O,{className:"text-xl text-js"})}),e.jsx(B,{mode:"wait",children:s&&e.jsx(m.span,{variants:rs,initial:"collapsed",animate:"expanded",exit:"collapsed",className:"font-bold text-js font-space",children:"CodeLinesJS"})})]})}));Is.displayName="NavigationLogo";const ns=r.memo(({icon:s,label:t,isActive:a=!1,isExpanded:i,onClick:n,variant:o="default",id:c})=>e.jsxs(m.button,{onClick:n,"data-testid":`nav-button-${t==null?void 0:t.toLowerCase().replace(/\s/g,"-")}`,className:`
       ${c==="code"&&"hidden lg:flex"}
       w-full flex items-center gap-3 px-4 py-3 
       rounded-lg transition-all duration-200
@@ -47,7 +47,7 @@ var Ei=Object.defineProperty,Ci=Object.defineProperties;var Ai=Object.getOwnProp
     max-w-[1000px]
     mx-auto
     relative
-  `,actions:"flex items-center gap-3 absolute right-6",gradients:{text:"bg-gradient-to-r from-js to-js/80 bg-clip-text text-transparent",hover:"hover:bg-js/5",base:"bg-gradient-to-br from-js/5 to-js/10"},borders:{base:"border-js/20",hover:"hover:border-js/30"},transitions:{base:"transition-all duration-200"},text:{heading:"text-lg font-bold text-js",primary:"text-white",secondary:"text-gray-400"}},Ir=()=>r.useMemo(()=>({container:{initial:{y:-20,opacity:0},animate:{y:0,opacity:1,transition:{duration:.3,ease:"easeOut"}}}}),[]),Br=s=>L(void 0,null,function*(){const t=yield fetch(`${A}users/profile`,{headers:{Authorization:`Bearer ${s}`}});if(!t.ok)throw new Error("Nie udało się pobrać danych użytkownika");return t.json()}),Mr=["userProfile"],Gr=()=>{const{token:s}=F();return G({queryKey:Mr,queryFn:()=>Br(s)})},Qs=r.memo(()=>{const s=Ir(),{data:t,isLoading:a,error:i}=Gr(),n=r.useMemo(()=>i?"Użytkowniku":a?"Ładowanie...":(t==null?void 0:t.data.user.username)||"Użytkowniku",[t==null?void 0:t.data.user.username,a,i]);return e.jsx(is,{children:e.jsx(m.nav,{variants:s.container,initial:"initial",animate:"animate",className:De.container,"aria-label":"Górna nawigacja",children:e.jsxs("div",{className:De.content,children:[e.jsx(Hs,{username:n}),e.jsx("div",{className:De.actions,children:e.jsx(r.Suspense,{fallback:e.jsx("div",{className:"w-10 h-10 rounded-full bg-gray-700 animate-pulse"}),children:e.jsx(qs,{})})})]})})})});Qs.displayName="TopNavigation";const Hr=r.lazy(()=>K(()=>import("./app-components-DHLBCYYN.js").then(s=>s.D),__vite__mapDeps([0,1,2,3,4,5,6]))),Ur=r.lazy(()=>K(()=>import("./app-components-DHLBCYYN.js").then(s=>s.q),__vite__mapDeps([0,1,2,3,4,5,6])).then(s=>({default:s.LoadingSpinner}))),Ks=r.memo(()=>{const s=Ms(),t=Rs(),{logout:a}=F();return e.jsx(is,{fallback:e.jsx(r.Suspense,{fallback:e.jsx("div",{className:"flex items-center justify-center min-h-screen",children:"Ładowanie..."}),children:e.jsx(Hr,{})}),children:e.jsxs("main",{className:"bg-gradient-to-b from-dark via-dark-medium to-dark backdrop-blur-lg flex min-h-screen",children:[e.jsxs(Y,{children:[e.jsx("title",{children:"Dashboard | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Panel użytkownika CodeLinesJS - dostęp do kursów, gier i statystyk nauki."})]}),e.jsx(Gs,{}),e.jsxs("div",{className:`flex-1 ${!s&&!t?"ml-[100px]":""} md:ml-[100px]`,children:[e.jsx(Qs,{}),e.jsx("div",{className:"mt-20 p-6",children:e.jsxs("div",{className:`${s||t?"flex":""} flex-col justify-center items-center gap-4 md:block`,children:[s||t?e.jsxs("div",{className:"flex flex-col gap-2 mb-4 md:hidden",children:[e.jsx(fe,{to:"/dashboard",children:e.jsx("button",{"aria-label":"Wróć do strony głównej",className:"bg-js text-dark px-4 py-2 rounded-md w-full flex items-center justify-center",children:"Wróć do Home"})}),e.jsxs("button",{onClick:a,"aria-label":"Wyloguj się",className:"bg-red-600 text-white px-4 py-2 rounded-md flex items-center justify-center gap-2",children:[e.jsx(bs,{"aria-hidden":"true"}),"Wyloguj się"]})]}):null,e.jsx(r.Suspense,{fallback:e.jsx(Ur,{text:"Ładowanie zawartości..."}),children:e.jsx(Ii,{})})]})})]})]})})});Ks.displayName="Dashboard";const cl=Object.freeze(Object.defineProperty({__proto__:null,default:Ks},Symbol.toStringTag,{value:"Module"})),pe={text:{heading:"font-bold text-js"},grid:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8",card:{base:"bg-dark/50 backdrop-blur-sm border border-js/10 shadow-xl",community:`
+  `,actions:"flex items-center gap-3 absolute right-6",gradients:{text:"bg-gradient-to-r from-js to-js/80 bg-clip-text text-transparent",hover:"hover:bg-js/5",base:"bg-gradient-to-br from-js/5 to-js/10"},borders:{base:"border-js/20",hover:"hover:border-js/30"},transitions:{base:"transition-all duration-200"},text:{heading:"text-lg font-bold text-js",primary:"text-white",secondary:"text-gray-400"}},Ir=()=>r.useMemo(()=>({container:{initial:{y:-20,opacity:0},animate:{y:0,opacity:1,transition:{duration:.3,ease:"easeOut"}}}}),[]),Br=s=>L(void 0,null,function*(){const t=yield fetch(`${A}users/profile`,{headers:{Authorization:`Bearer ${s}`}});if(!t.ok)throw new Error("Nie udało się pobrać danych użytkownika");return t.json()}),Mr=["userProfile"],Gr=()=>{const{token:s}=F();return G({queryKey:Mr,queryFn:()=>Br(s)})},Qs=r.memo(()=>{const s=Ir(),{data:t,isLoading:a,error:i}=Gr(),n=r.useMemo(()=>i?"Użytkowniku":a?"Ładowanie...":(t==null?void 0:t.data.user.username)||"Użytkowniku",[t==null?void 0:t.data.user.username,a,i]);return e.jsx(is,{children:e.jsx(m.nav,{variants:s.container,initial:"initial",animate:"animate",className:De.container,"aria-label":"Górna nawigacja",children:e.jsxs("div",{className:De.content,children:[e.jsx(Hs,{username:n}),e.jsx("div",{className:De.actions,children:e.jsx(r.Suspense,{fallback:e.jsx("div",{className:"w-10 h-10 rounded-full bg-gray-700 animate-pulse"}),children:e.jsx(qs,{})})})]})})})});Qs.displayName="TopNavigation";const Hr=r.lazy(()=>K(()=>import("./app-components-DiGU-x_9.js").then(s=>s.D),__vite__mapDeps([0,1,2,3,4,5,6]))),Ur=r.lazy(()=>K(()=>import("./app-components-DiGU-x_9.js").then(s=>s.q),__vite__mapDeps([0,1,2,3,4,5,6])).then(s=>({default:s.LoadingSpinner}))),Ks=r.memo(()=>{const s=Ms(),t=Rs(),{logout:a}=F();return e.jsx(is,{fallback:e.jsx(r.Suspense,{fallback:e.jsx("div",{className:"flex items-center justify-center min-h-screen",children:"Ładowanie..."}),children:e.jsx(Hr,{})}),children:e.jsxs("main",{className:"bg-gradient-to-b from-dark via-dark-medium to-dark backdrop-blur-lg flex min-h-screen",children:[e.jsxs(Y,{children:[e.jsx("title",{children:"Dashboard | CodeLinesJS"}),e.jsx("meta",{name:"description",content:"Panel użytkownika CodeLinesJS - dostęp do kursów, gier i statystyk nauki."})]}),e.jsx(Gs,{}),e.jsxs("div",{className:`flex-1 ${!s&&!t?"ml-[100px]":""} md:ml-[100px]`,children:[e.jsx(Qs,{}),e.jsx("div",{className:"mt-20 p-6",children:e.jsxs("div",{className:`${s||t?"flex":""} flex-col justify-center items-center gap-4 md:block`,children:[s||t?e.jsxs("div",{className:"flex flex-col gap-2 mb-4 md:hidden",children:[e.jsx(fe,{to:"/dashboard",children:e.jsx("button",{"aria-label":"Wróć do strony głównej",className:"bg-js text-dark px-4 py-2 rounded-md w-full flex items-center justify-center",children:"Wróć do Home"})}),e.jsxs("button",{onClick:a,"aria-label":"Wyloguj się",className:"bg-red-600 text-white px-4 py-2 rounded-md flex items-center justify-center gap-2",children:[e.jsx(bs,{"aria-hidden":"true"}),"Wyloguj się"]})]}):null,e.jsx(r.Suspense,{fallback:e.jsx(Ur,{text:"Ładowanie zawartości..."}),children:e.jsx(Ii,{})})]})})]})]})})});Ks.displayName="Dashboard";const cl=Object.freeze(Object.defineProperty({__proto__:null,default:Ks},Symbol.toStringTag,{value:"Module"})),pe={text:{heading:"font-bold text-js"},grid:"grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8",card:{base:"bg-dark/50 backdrop-blur-sm border border-js/10 shadow-xl",community:`
       col-span-1 md:col-span-2 lg:col-span-1
       order-2 lg:order-1
     `,game:`
