--- server/public/assets/app-services-tJo8l8LQ.js
@@ -1,1 +1,0 @@
-var S=Object.defineProperty,T=Object.defineProperties;var E=Object.getOwnPropertyDescriptors;var z=Object.getOwnPropertySymbols;var A=Object.prototype.hasOwnProperty,B=Object.prototype.propertyIsEnumerable;var k=(t,e,s)=>e in t?S(t,e,{enumerable:!0,configurable:!0,writable:!0,value:s}):t[e]=s,i=(t,e)=>{for(var s in e||(e={}))A.call(e,s)&&k(t,s,e[s]);if(z)for(var s of z(e))B.call(e,s)&&k(t,s,e[s]);return t},p=(t,e)=>T(t,E(e));var u=(t,e,s)=>new Promise((w,f)=>{var c=o=>{try{d(s.next(o))}catch(m){f(m)}},h=o=>{try{d(s.throw(o))}catch(m){f(m)}},d=o=>o.done?w(o.value):Promise.resolve(o.value).then(c,h);d((s=s.apply(t,e)).next())});import{A as N}from"./app-auth-DYLPzh_K.js";const P=t=>{const e=t.message.toLowerCase();return e.includes("failed to fetch")?"Nie można połączyć się z serwerem. Sprawdź swoje połączenie internetowe.":e.includes("network request failed")?"Błąd połączenia sieciowego. Sprawdź swoje połączenie internetowe.":e.includes("timeout")?"Przekroczono czas oczekiwania na odpowiedź serwera.":e.includes("abort")?"Żądanie zostało przerwane.":e.includes("cors")?"Błąd polityki CORS. Brak dostępu do zasobu z tej domeny.":t.message},j={request(s){return u(this,arguments,function*(t,e={}){var w,f;try{const{method:c="GET",headers:h={},body:d,requiresAuth:o=!1}=e,q=`${N.replace("www.","")}${t}`,y=i({"Content-Type":"application/json",Accept:"application/json"},h);if(o){const l=localStorage.getItem("token")||sessionStorage.getItem("token");l&&(y.Authorization=`Bearer ${l}`)}const g={method:c,headers:y,mode:"cors",credentials:"include"};d&&(g.body=JSON.stringify(d));const a=yield fetch(q,g);let r=null,n=null;try{const l=yield a.text();l&&(r=JSON.parse(l))}catch(l){n="Nieprawidłowa odpowiedź serwera"}if(r){if(r.status==="success")return{data:r.data,error:null,status:r.code||a.status};if(r.status==="fail"||r.status==="error")return n=r.message||((f=(w=r.errors)==null?void 0:w[0])==null?void 0:f.message)||`Błąd: ${r.status}`,{data:null,error:n,status:r.code||a.status};if(r.error)return{data:null,error:r.error,status:a.status};if(r.token)return{data:r,error:null,status:a.status}}return!a.ok&&!n?(n=`Błąd ${a.status}: ${a.statusText}`,{data:null,error:n,status:a.status}):{data:(r==null?void 0:r.data)||r,error:n,status:a.status}}catch(c){return{data:null,error:c instanceof Error?P(c):"Nieznany błąd zapytania",status:0}}})},get(s){return u(this,arguments,function*(t,e={}){return this.request(t,p(i({},e),{method:"GET"}))})},post(w,f){return u(this,arguments,function*(t,e,s={}){return this.request(t,p(i({},s),{method:"POST",body:e}))})},put(w,f){return u(this,arguments,function*(t,e,s={}){return this.request(t,p(i({},s),{method:"PUT",body:e}))})},delete(s){return u(this,arguments,function*(t,e={}){return this.request(t,p(i({},e),{method:"DELETE"}))})},patch(w,f){return u(this,arguments,function*(t,e,s={}){return this.request(t,p(i({},s),{method:"PATCH",body:e}))})}};export{j as h};
