--- server/public/assets/index-BQhTrSRm.js
@@ -1,2 +1,0 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/app-dashboard-BluXOO0L.js","assets/app-auth-DYLPzh_K.js","assets/vendor-react-RXQ05gJ2.js","assets/vendor-other-C1_ffh8T.js","assets/vendor-monaco-Dk8M9SDd.js","assets/app-components-DHLBCYYN.js","assets/vendor-framer-motion-SZOJYst1.js","assets/app-services-tJo8l8LQ.js","assets/vendor-charts-Czvt6lRK.js","assets/page-home.tsx-CdAWi1XM.js","assets/page-auth.tsx-dG67t_Qf.js","assets/page-privacypolicy.tsx-CIBx5sMh.js"])))=>i.map(i=>d[i]);
-import{r as n,aP as _,j as t,aQ as p,aR as h,aS as m,aT as f,aU as E,aV as x}from"./vendor-react-RXQ05gJ2.js";import{ax as y}from"./vendor-other-C1_ffh8T.js";import{_ as o}from"./app-auth-DYLPzh_K.js";import{L as j}from"./app-components-DHLBCYYN.js";import"./vendor-monaco-Dk8M9SDd.js";import"./vendor-framer-motion-SZOJYst1.js";import"./app-services-tJo8l8LQ.js";(function(){const a=document.createElement("link").relList;if(a&&a.supports&&a.supports("modulepreload"))return;for(const s of document.querySelectorAll('link[rel="modulepreload"]'))l(s);new MutationObserver(s=>{for(const i of s)if(i.type==="childList")for(const d of i.addedNodes)d.tagName==="LINK"&&d.rel==="modulepreload"&&l(d)}).observe(document,{childList:!0,subtree:!0});function u(s){const i={};return s.integrity&&(i.integrity=s.integrity),s.referrerPolicy&&(i.referrerPolicy=s.referrerPolicy),s.crossOrigin==="use-credentials"?i.credentials="include":s.crossOrigin==="anonymous"?i.credentials="omit":i.credentials="same-origin",i}function l(s){if(s.ep)return;s.ep=!0;const i=u(s);fetch(s.href,i)}})();const P=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.C),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.CommunitySection}))),c=n.lazy(()=>o(()=>import("./app-components-DHLBCYYN.js").then(e=>e.r),__vite__mapDeps([5,2,3,4,6,1,7])).then(e=>({default:e.ErrorPage}))),g=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.D),__vite__mapDeps([0,1,2,3,4,5,6,7,8]))),L=n.lazy(()=>o(()=>import("./app-components-DHLBCYYN.js").then(e=>e.s),__vite__mapDeps([5,2,3,4,6,1,7])).then(e=>({default:e.ProtectedRoute}))),R=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.a),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.DashboardContent}))),O=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.S),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.StatsSection}))),D=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.L),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.LearnSection}))),S=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.b),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.LessonPage}))),I=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.c),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.SettingsSection}))),A=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.G),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.GamesSection}))),T=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.d),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.GameplayRouter}))),z=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.e),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.CodeEditor}))),V=n.lazy(()=>o(()=>import("./app-auth-DYLPzh_K.js").then(e=>e.i),__vite__mapDeps([1,2,3,4,5,6,7])).then(e=>({default:e.ResetPasswordPage}))),b=n.lazy(()=>o(()=>import("./app-dashboard-BluXOO0L.js").then(e=>e.R),__vite__mapDeps([0,1,2,3,4,5,6,7,8])).then(e=>({default:e.ReportBugPage}))),C=n.lazy(()=>o(()=>import("./page-home.tsx-CdAWi1XM.js"),__vite__mapDeps([9,2,3,4,5,6,1,7]))),w=n.lazy(()=>o(()=>import("./page-auth.tsx-dG67t_Qf.js"),__vite__mapDeps([10,2,3,4,5,6,1,7]))),G=n.lazy(()=>o(()=>import("./page-privacypolicy.tsx-CIBx5sMh.js"),__vite__mapDeps([11,2,3,4,5,6,1,7]))),r=({children:e})=>t.jsx(n.Suspense,{fallback:t.jsx(j,{fullScreen:!0,text:"Ładowanie strony..."}),children:e}),v=_([{path:"/",element:t.jsx(r,{children:t.jsx(C,{})}),errorElement:t.jsx(r,{children:t.jsx(c,{})})},{path:"/logowanie",element:t.jsx(r,{children:t.jsx(w,{})})},{path:"/reset-password/:token",element:t.jsx(r,{children:t.jsx(V,{})})},{path:"/polityka-prywatnosci",element:t.jsx(r,{children:t.jsx(G,{})})},{path:"/dashboard",element:t.jsx(r,{children:t.jsx(L,{children:t.jsx(g,{})})}),children:[{index:!0,element:t.jsx(r,{children:t.jsx(R,{})})},{path:"stats",element:t.jsx(r,{children:t.jsx(O,{})})},{path:"community/*",element:t.jsx(r,{children:t.jsx(P,{})})},{path:"learn",element:t.jsx(r,{children:t.jsx(D,{})})},{path:"learn/lesson/:lessonSlug",element:t.jsx(r,{children:t.jsx(S,{})})},{path:"settings",element:t.jsx(r,{children:t.jsx(I,{})})},{path:"play",element:t.jsx(r,{children:t.jsx(A,{})})},{path:"play/:slug",element:t.jsx(r,{children:t.jsx(T,{})})},{path:"code",element:t.jsx(r,{children:t.jsx(z,{})})},{path:"report",element:t.jsx(r,{children:t.jsx(b,{})})}]},{path:"*",element:t.jsx(r,{children:t.jsx(c,{})})}]),q="466348895611-kdgecrm842jsilesaj0s0c5bnqn7maiq.apps.googleusercontent.com",B=new y;function N(){return t.jsxs(p,{client:B,children:[t.jsx(h,{position:"top-right",toastOptions:{duration:3e3,style:{background:"#1E1E1E",color:"#fff",border:"1px solid rgba(247, 223, 30, 0.1)"},success:{iconTheme:{primary:"#F7DF1E",secondary:"#1E1E1E"}},error:{iconTheme:{primary:"#ff4b4b",secondary:"#1E1E1E"}}}}),t.jsx(m,{clientId:q,onScriptLoadError:()=>console.error("Google Script failed to load"),children:t.jsx(f,{router:v})})]})}E.createRoot(document.getElementById("root")).render(t.jsx(n.StrictMode,{children:t.jsx(x,{children:t.jsx(N,{})})}));
