+++ server/public/assets/page-auth.tsx-DjcQ4rlt.js
@@ -1,0 +1,1 @@
+import{j as r,r as t}from"./vendor-react-RXQ05gJ2.js";import{f as o}from"./app-components-DiGU-x_9.js";import{b as s}from"./app-auth-CQDHbRjF.js";import"./vendor-other-C1_ffh8T.js";import"./vendor-monaco-Dk8M9SDd.js";import"./vendor-framer-motion-SZOJYst1.js";import"./app-services-BaprE51I.js";const c=()=>r.jsx(t.Suspense,{fallback:r.jsx(o,{}),children:r.jsx(s,{})});export{c as default};
