--- server/public/assets/page-auth.tsx-dG67t_Qf.js
@@ -1,1 +1,0 @@
-import{j as r,r as t}from"./vendor-react-RXQ05gJ2.js";import{f as o}from"./app-components-DHLBCYYN.js";import{b as s}from"./app-auth-DYLPzh_K.js";import"./vendor-other-C1_ffh8T.js";import"./vendor-monaco-Dk8M9SDd.js";import"./vendor-framer-motion-SZOJYst1.js";import"./app-services-tJo8l8LQ.js";const c=()=>r.jsx(t.Suspense,{fallback:r.jsx(o,{}),children:r.jsx(s,{})});export{c as default};
