--- server/public/assets/page-home.tsx-CdAWi1XM.js
@@ -1,1 +1,0 @@
-import{j as s}from"./vendor-react-RXQ05gJ2.js";import{H as o,i as t,j as e,k as a,T as i,l as r,W as m,m as j,n as x,o as n,p as c}from"./app-components-DHLBCYYN.js";import"./vendor-other-C1_ffh8T.js";import"./vendor-monaco-Dk8M9SDd.js";import"./vendor-framer-motion-SZOJYst1.js";import"./app-auth-DYLPzh_K.js";import"./app-services-tJo8l8LQ.js";const H=()=>s.jsxs("div",{className:"bg-gray-900",children:[s.jsx(o,{}),s.jsx(t,{}),s.jsx(e,{}),s.jsx(a,{}),s.jsx(i,{}),s.jsx(r,{}),s.jsx(m,{}),s.jsx(j,{}),s.jsx(x,{}),s.jsx(n,{}),s.jsx(c,{})]});export{H as default};
