+++ server/public/index.html
@@ -9,15 +9,15 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@400;500;600;700&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=JetBrains+Mono:wght@400;500&display=swap" rel="stylesheet">
-    <script type="module" crossorigin src="/assets/index-BQhTrSRm.js"></script>
+    <script type="module" crossorigin src="/assets/index-DMx6TykQ.js"></script>
     <link rel="modulepreload" crossorigin href="/assets/vendor-other-C1_ffh8T.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-monaco-Dk8M9SDd.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-react-RXQ05gJ2.js">
     <link rel="modulepreload" crossorigin href="/assets/vendor-framer-motion-SZOJYst1.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-components-DHLBCYYN.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-services-tJo8l8LQ.js">
-    <link rel="modulepreload" crossorigin href="/assets/app-auth-DYLPzh_K.js">
-    <link rel="stylesheet" crossorigin href="/assets/index-BBca1lE6.css">
+    <link rel="modulepreload" crossorigin href="/assets/app-components-DiGU-x_9.js">
+    <link rel="modulepreload" crossorigin href="/assets/app-services-BaprE51I.js">
+    <link rel="modulepreload" crossorigin href="/assets/app-auth-CQDHbRjF.js">
+    <link rel="stylesheet" crossorigin href="/assets/index-BFzPeBr2.css">
   </head>
   <body>
     <div id="root"></div>
